backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyspur"
version = "0.0.21"
description = "PySpur is a Python library for building and deploying AI Agents."
requires-python = ">=3.12"
license = "Apache-2.0"
maintainers = [
    {name = "Srijan Patel", email = "srijan@pyspur.dev"},
    {name = "Jean Kaddour", email = "jean@pyspur.dev"},
    {name = "Parshva Bhadra", email = "parshva.bhadra@pyspur.dev"},
]
dependencies = [
    "alembic==1.14.0",
    "arrow==1.3.0",
    "asyncio==3.4.3",
    "attrs==24.3.0",
    "backend==0.2.4.1",
    "chromadb==0.6.2",
    "datasets==3.2.0",
    "docx2txt==0.8",
    "docx2python==3.3.0",
    "fastapi==0.115.6",
    "genanki==0.13.1",
    "google-api-python-client==2.159.0",
    "grpcio==1.69.0",
    "Jinja2==3.1.5",
    "litellm==1.57.5",
    "loguru==0.7.3",
    "numpy==2.2.1",
    "ollama==0.4.5",
    "pandas==2.2.3",
    "pinecone==5.4.2",
    "praw==7.8.1",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.5",
    "pypdf==5.1.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "python-pptx==1.0.2",
    "PyYAML==6.0.2",
    "py-zerox==0.0.7",
    "qdrant_client==1.12.2",
    "redis==5.2.1",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-file==2.1.0",
    "requests-oauthlib==1.3.1",
    "retrying==1.3.4",
    "slack_sdk==3.34.0",
    "SQLAlchemy==2.0.36",
    "supabase==2.11.0",
    "six==1.17.0",
    "tenacity==8.3.0",
    "tiktoken==0.7.0",
    "tqdm==4.67.1",
    "weaviate_client==4.10.2",
    "itsdangerous==2.2.0",
    "phidata==2.7.8",
    "youtube_transcript_api==0.6.3",
    "PyGithub==2.5.0",
    "firecrawl-py==1.10.2",
    "httpx[http2]==0.27.2",
    "sendgrid==6.11.0",
    "resend==2.6.0",
]

[project.urls]
Repository = "https://github.com/pyspur-dev/pyspur"
Documentation = "https://docs.pyspur.dev"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
universal = false
packages = ["pyspur"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C", "D"]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in __init__
    "D203",  # Ignore 'incorrect-blank-line-before-class' in favor of D211
    "D213",  # Ignore 'multi-line-summary-second-line' in favor of D212
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=pyspur --cov-report=term-missing" 